crime-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
uvloop==0.19.0
httptools==0.6.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
email_validator
passlib[bcrypt]==1.7.4
shortuuid==1.0.13
python-jose[cryptography]
redis==5.2.1 
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.4.0 
httpx==0.26.0
googlemaps
folium==0.14.0